separate build and champion folders
