+++ src/components/builds/utils.js
@@ -36,8 +36,6 @@ export const Header = (title) => {
 export const Winrate = (winrate) => {
   const classes = useStyles();
   return (
-    <div className={classes.itemWinrate}>
-      {winrate.toFixed(1) + '% WR'}
-    </div>
+    <div className={classes.itemWinrate}>{winrate.toFixed(1) + '% WR'}</div>
   );
 };
