+++ src/index.js
@@ -10,7 +10,7 @@ import Leaderboards from './components/leaderboard/leaderboard_page.js';
 import About from './components/about/about_page.js';
 import Faq from './components/faq/faq_page.js';
 import ChampionsPage from './components/champions/champions_page.js';
-import Champion from './components/champions/champion.js';
+import Champion from './components/builds/champion.js';
 import TierlistPage from './components/tierlist/tierlist_page.js';
 import { Container } from '@material-ui/core';
 import './stylesheets/style.css';
